+++ examples/multiple_buckets/main.tf
@@ -1,0 +1,39 @@
+/**
+ * Copyright 2018 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+provider "google" {
+  version = "~> 2.18.0"
+}
+
+module "cloud_storage" {
+  source             = "../.."
+  project_id         = var.project_id
+  prefix             = var.prefix
+  names              = var.names
+  bucket_policy_only = var.bucket_policy_only
+
+  lifecycle_rules = [{
+    action = {
+      type          = "SetStorageClass"
+      storage_class = "NEARLINE"
+    }
+    condition = {
+      age                   = "10"
+      matches_storage_class = "MULTI_REGIONAL,STANDARD,DURABLE_REDUCED_AVAILABILITY"
+    }
+  }]
+}
+